requirements.txt
pelican[markdown]
pelican-seo
pelican-jinja2content
pylast
typogrify
Pillow
python-dotenv
